Hand Written Digits Classification using Neural N.
Overview:
This project is a simple yet fundamental deep learning task that demonstrates the power of neural networks in digits classification. It focuses on recognizing and classifying handwritten digits from 0 to 9. The goal is to build a neural network model that can accurately identify the digit represented in a given image.
Features:
Dataset: We used the popular keras dataset, which contains a large collection of handwritten digits with corresponding labels for training and testing purposes.

Model Architecture: The neural network architecture consists of layers of densely connected neurons, providing a foundation for digit feature extraction and classification.

Training: The model was trained using various optimization techniques and hyperparameter tuning to achieve high accuracy in digit recognition.

Evaluation: We used metrics such as accuracy, confusion matrix, and loss curves to assess the model's performance on both the training and testing datasets.